refactor: Genre class import refactor
