+++ src/main/java/com/northcoders/recordshopbackend/model/dto/GetAlbumDTO.java
@@ -1,5 +1,5 @@
 package com.northcoders.recordshopbackend.model.dto;
 
-import com.northcoders.recordshopbackend.model.Genre;
+import com.northcoders.recordshopbackend.model.enums.Genre;
 
 public record GetAlbumDTO(Long id, String title, String artist, Genre genre, String releaseDate, int stock) {}
